move class definitions to headers
